Change star rating response data type.

+++ server/SvyU.Models/StarRateResponse.cs
@@ -7,6 +7,6 @@ namespace SvyU.Models
         public QuestionType Type => QuestionType.StarRate;
 
         [JsonProperty("answer")]
-        public int Response { get; set; }
+        public double Response { get; set; }
     }
 }
